+++ kernel/src/memory/mod.rs
@@ -2,6 +2,10 @@
 
 use core::cmp::{max, min};
 
+mod physical;
+
+pub use physical::{AllocationError, FrameAllocator, FrameRange, MemoryMap, MemoryMapError};
+
 /// Phoenix's base page size in bytes.
 pub const PAGE_SIZE: usize = 4096;